+++ src/main/java/site/matzip/comment/repository/CommentRepository.java
@@ -1,6 +1,7 @@
-package site.matzip.comment.domain;
+package site.matzip.comment.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import site.matzip.comment.domain.Comment;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
 }
