+++ src/test/java/site/matzip/comment/domain/service/CommentServiceTest.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.transaction.annotation.Transactional;
+import site.matzip.comment.service.CommentService;
 
 @SpringBootTest
 @Transactional
